Using unsigned integer insted of floats. Having some problems with blending for uints.

+++ shaders/paralell.frag
@@ -1,8 +1,9 @@
 #version 330
 
-out vec4 out_Color;
+//out uint out_Color;
+out uvec4 out_Color;
 
-//set white color with full opacity
 void main(void){
-	out_Color = vec4(1,0,0,1);
+	//out_Color = uint(1);
+	out_Color = uvec4(1, 0, 0, 1);
 }
